Refactor interpreter evaluation
Before interpreter evaluation was done at run-time when the interpreter
was first used.

With this patch interpreters are looked up when the environments are
generated.

Also more precise interpreter specification is possible. Before,
specifying "3" in a riotfile would result in a "python3" executable
being looked up. So riot would fail if Python was installed as "python"
and no "python3" executable was available on the path.

Now the path is searched for python executables which are then
referenced when resolving the python to use for an environment.