fix: install layer dependencies only
With the idea of adding to the PYTHONPATH, each venv layer only needs to
install its packages inside the venv instance, thus reducing the number
of packages that are installed in each venv instance. We still use the
full package spec in the result report for easier identification of all
the dependencies used.

BREAKING CHANGE: this change implies that package declarations that are
used for pinning dependencies will have to be added to each venv layer
that requires them.