feat: add `create` argument to Venv (#149)
* perf: create base venvs only

With control over the pip installation prefix there is no need to create
venvs other than the base one. This saves a lot of memory and speeds up
the task runs.

* fix: install dependencies only if prefix does not exist

* feat: add create argument to Venv

The new argument allows explicitly creating a virtual environment
instance with all the dependencies installed within it rather than in
separate prefixes. This is useful for packages that do not support
dependencies installed in other prefixes.