Quote paths that might contain special chars (#151)
Some the the virtualenv paths generated may include shell special
characters like `[` and `]`. Namely this happens when extras are
provided (eg. `reno[sphinx]`).

This was breaking the virtualenv commands as the shell attempts to
interpret the special characters in the commands. This is fixed by
quoting the strings before we pass them to the shell.