fix: add quotes around cmdargs
This fixes use cases such as:

  riot run foobar -- -k 'not test_foobaz'

for:

  pytest {cmdargs}

Which is broken right now because it turns out to run:

as:

  pytest -k not test_foobaz

This fixes the code by adding '' around each cmdargs item, which turns the
command line into:

  pytest '-k' 'not test_foobaz'

Since this whole string is passed to subprocess with shell=True, we're sure
that shell does not try to interpolate anything.

That sounds safe and neat.