refactor: make Venv a simple class
The release 0.960 of mypy has highlighted a misuse of `InitVar`. This
change fixes it by removing the `dataclass` decorator from `Venv` and
adding the required boilerplate code for the `__init__` method. The
different `__repr__` method that follows from this refactor has caused
the short hashes to change and the tests had to be adjusted accordingly.