refactor!: ensure dev package is installed in base venv (#183)
When running a command for the first time with the -s option, the base
virtual environments get created, but the installation of the dev
package is skipped. This results in import errors when running Python
commands against the dev package. This change modified the behaviour of
riot to ensure that the dev package is always installed when the base
virtual environments are generated for the first time. If no dev package
is present in the working directory, a warning is issued instead of an
error to allow running generic commands without the use of the -s
option.