fix: install dev package in dedicated venv
We install the dev package in its own virtual environment. We use
a pristine base venv just for installing dependencies in dependency
layers. We then use the venv with the dev package to run commands
and spawn shells. This way we prevent dependency layer from
removing packages from the base venv with the dev package.