refactor: split into submodule and use more modern Python
We split the big riot.py monolith into smaller submodules. This
helps modularise the code base and ensure that each unit takes
care of just its responsibility.

The codebase has also been improved with the use of pathlib
replacing usages of os.path. The change also adds envier for
parsing the configuration from the environment, and an isort
configuration.

We also drop support for 3.7.